pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "academic-insight-ai"
version = "0.1.0"
description = "Academic local LLM workbench for repeatable task pipelines"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.7,<3",
  "python-dotenv>=1.0,<2",
  "requests>=2.32,<3",
  "typer>=0.12,<1",
]

[project.scripts]
academic-ai = "academic_insight_ai.cli.main:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
